+++ NEWS.md
@@ -28,6 +28,7 @@
 * Created `release` function and S3 `plot` method for merging multiple schemes (i.e., blocks or phases) into one release. This is used to inspect how multiple back-to-back phases mold together
 * Added `perc_1RM` argument to `scheme_manual()` for the user to provide manual 1RM percentages, rather than to be estimated
 * Added `scheme_perc_1RM()` which is simpler `scheme_manual()` for manually entering 1RM percentages. For example creating simple warm-up scheme 
+* Added `+` method for `STMr_scheme` objects. This allows for easy modular adding of the schemes
 
 # STMr 0.1.3